chore: move payout function into own file
